Fix some bugs; decrease required cmake version to 3.10; move game struct to a separate file; auto-create games dir
